# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vertex Finance OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: accountId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    put:
      operationId: updateTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /transactions/{id}/duplicate:
    post:
      operationId: duplicateTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── MONTHLY PLANS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        competenceDate:
          type: string
          format: date
        movementDate:
          type: string
          format: date
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        subcategoryId:
          type: integer
          nullable: true
        subcategoryName:
          type: string
          nullable: true
        type:
          type: string
          enum: [income, expense, transfer]
        paymentMethod:
          type: string
          nullable: true
        description:
          type: string
        amount:
          type: number
        accountId:
          type: integer
          nullable: true
        accountName:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, paid, received]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, competenceDate, movementDate, type, description, amount, status, createdAt]
    CreateTransactionInput:
      type: object
      properties:
        competenceDate:
          type: string
          format: date
        movementDate:
          type: string
          format: date
        categoryId:
          type: integer
          nullable: true
        subcategoryId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [income, expense, transfer]
        paymentMethod:
          type: string
          nullable: true
        description:
          type: string
        amount:
          type: number
        accountId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [planned, paid, received]
        notes:
          type: string
          nullable: true
      required: [competenceDate, movementDate, type, description, amount, status]
...